Malloc Free

+++ 0x0B-malloc_free/3-alloc_grid.c
@@ -13,14 +13,14 @@ int i, j = 0;
 if (width <= 0 || height <= 0)
 return (NULL);
 
-g = malloc(height * sizeof(*g));
+g = (int) malloc(height * sizeof(*g));
 
 if (g == NULL)
 return (NULL);
 
 for (i = 0; i < height; i++)
 {
-g[i] = malloc(width * sizeof(**g));
+g[i] = (int) malloc(width * sizeof(**g));
 if (g[i] == NULL)
 {
 free(g);
